comprendre/requirements.txt
Django==3.1.3
psycopg2-binary==2.8.6
django-nose==1.4.7
coverage==5.3
pylint==2.6.0
pylint-django==2.3.0
django-taggit==1.3.0
djangorestframework==3.12.2
django-filter==2.4.0
drf-yasg==1.20.0
